requirements.txt
pyserial
wave
